update to laravel 6.0

+++ app/Parse_mods/defense_stats/Evasion.php
@@ -19,7 +19,7 @@ class Evasion extends Stat
     {
         $this->setType($type);
         $this->name ='Evasion';
-        $modValue = filter_var($mod, FILTER_SANITIZE_NUMBER_INT);
+        $modValue = (int) filter_var($mod, FILTER_SANITIZE_NUMBER_INT);
         $modName = preg_replace('/\d+/u', '#', $mod);
 
         // From Base_Stats.php
