+++ app/Parse_mods/defense_stats/Percentage_Mana.php
@@ -14,7 +14,7 @@ class Percentage_Mana extends Stat
     {
         $this->setType($type);
         $this->name ='% Mana';
-        $modValue = filter_var($mod, FILTER_SANITIZE_NUMBER_INT);
+        $modValue = (int) filter_var($mod, FILTER_SANITIZE_NUMBER_INT);
         $modName = preg_replace('/\d+/u', '#', $mod);
 
         if ( $modName === '#% increased maximum Mana') {
